fix: ensure unique edges
Only add a edge if it is unique. This allows for a lower run time but
also makes our graph design easier to understand and predict. It makes
no difference for us in functionality as we only care about what modules
we import and not what we import from those modules.

This solves https://github.com/benawad/destiny/issues/73.