fix: ensure path correctness
Previously we assumed that all args, that did not match any of our
defined options, was paths which should be restructured. This lead to
issues like #81 where a option was parsed as a path and the user isnt
informed.

Now the `cli` ensures that all paths is either a existing path or a
glob.